adk-agents/qa_agent/requirements.txt
google-adk
google-genai
playwright
pytest
requests
python-dotenv
